# specs/004-recipe-basic-crud/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Recipe Service API
  description: 원본 레시피 및 요리사 조회 API
  version: 1.0.0
  contact:
    name: Naecipe Backend Team

servers:
  - url: http://localhost:8001/api/v1
    description: Local development
  - url: https://api.naecipe.com/recipe/v1
    description: Production

tags:
  - name: recipes
    description: 레시피 관련 API
  - name: chefs
    description: 요리사 관련 API

paths:
  /recipes:
    get:
      tags:
        - recipes
      summary: 레시피 목록 조회
      description: 커서 기반 페이지네이션으로 레시피 목록을 조회합니다.
      operationId: listRecipes
      parameters:
        - name: cursor
          in: query
          description: 페이지네이션 커서 (이전 응답의 next_cursor)
          schema:
            type: string
        - name: limit
          in: query
          description: 페이지당 항목 수 (기본 20, 최대 100)
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeListResponse'
        '400':
          description: 잘못된 요청 (커서 형식 오류 등)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /recipes/{id}:
    get:
      tags:
        - recipes
      summary: 레시피 상세 조회
      description: 레시피 ID로 상세 정보를 조회합니다.
      operationId: getRecipe
      parameters:
        - name: id
          in: path
          required: true
          description: 레시피 UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeDetail'
        '404':
          description: 레시피를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /recipes/popular:
    get:
      tags:
        - recipes
      summary: 인기 레시피 조회
      description: 인기도 점수 기준으로 정렬된 레시피 목록을 조회합니다.
      operationId: getPopularRecipes
      parameters:
        - name: category
          in: query
          description: 카테고리 필터 (한식, 양식 등)
          schema:
            type: string
        - name: limit
          in: query
          description: 조회할 레시피 수 (기본 10, 최대 50)
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 50
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecipeListItem'

  /chefs:
    get:
      tags:
        - chefs
      summary: 요리사 목록 조회
      description: 커서 기반 페이지네이션으로 요리사 목록을 조회합니다.
      operationId: listChefs
      parameters:
        - name: cursor
          in: query
          description: 페이지네이션 커서
          schema:
            type: string
        - name: limit
          in: query
          description: 페이지당 항목 수 (기본 20, 최대 100)
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChefListResponse'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /chefs/{id}:
    get:
      tags:
        - chefs
      summary: 요리사 상세 조회
      description: 요리사 ID로 상세 정보를 조회합니다.
      operationId: getChef
      parameters:
        - name: id
          in: path
          required: true
          description: 요리사 UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChefDetail'
        '404':
          description: 요리사를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /chefs/{id}/recipes:
    get:
      tags:
        - chefs
      summary: 요리사별 레시피 조회
      description: 특정 요리사의 레시피 목록을 조회합니다.
      operationId: getChefRecipes
      parameters:
        - name: id
          in: path
          required: true
          description: 요리사 UUID
          schema:
            type: string
            format: uuid
        - name: cursor
          in: query
          description: 페이지네이션 커서
          schema:
            type: string
        - name: limit
          in: query
          description: 페이지당 항목 수 (기본 20, 최대 100)
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeListResponse'
        '404':
          description: 요리사를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /chefs/popular:
    get:
      tags:
        - chefs
      summary: 인기 요리사 조회
      description: 레시피 수, 조회수, 평점 기준으로 정렬된 요리사 목록을 조회합니다.
      operationId: getPopularChefs
      parameters:
        - name: limit
          in: query
          description: 조회할 요리사 수 (기본 10, 최대 50)
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 50
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChefListItem'

components:
  schemas:
    # --- Recipe Schemas ---
    RecipeListItem:
      type: object
      required:
        - id
        - title
        - cooking_time_minutes
        - difficulty
        - avg_rating
      properties:
        id:
          type: string
          format: uuid
          description: 레시피 UUID
        title:
          type: string
          description: 레시피 제목
          example: 김치찌개
        thumbnail_url:
          type: string
          nullable: true
          description: 썸네일 이미지 URL
        cooking_time_minutes:
          type: integer
          description: 조리 시간 (분)
          example: 30
        difficulty:
          type: string
          enum: [easy, medium, hard]
          description: 난이도
        avg_rating:
          type: number
          format: float
          description: 평균 평점 (0.0~5.0)
          example: 4.5
        chef_name:
          type: string
          nullable: true
          description: 요리사 이름
          example: 백종원

    RecipeDetail:
      allOf:
        - $ref: '#/components/schemas/RecipeListItem'
        - type: object
          required:
            - ingredients
            - steps
            - tags
          properties:
            description:
              type: string
              nullable: true
              description: 레시피 설명
            servings:
              type: integer
              nullable: true
              description: 인분 수
            video_url:
              type: string
              nullable: true
              description: 영상 URL
            view_count:
              type: integer
              description: 조회수
            save_count:
              type: integer
              description: 저장수
            ingredients:
              type: array
              items:
                $ref: '#/components/schemas/Ingredient'
            steps:
              type: array
              items:
                $ref: '#/components/schemas/CookingStep'
            tags:
              type: array
              items:
                $ref: '#/components/schemas/Tag'
            chef:
              $ref: '#/components/schemas/ChefSummary'
              nullable: true

    Ingredient:
      type: object
      required:
        - name
        - order_index
      properties:
        name:
          type: string
          description: 재료명
          example: 돼지고기
        amount:
          type: string
          nullable: true
          description: 양
          example: "200"
        unit:
          type: string
          nullable: true
          description: 단위
          example: g
        order_index:
          type: integer
          description: 표시 순서
        is_optional:
          type: boolean
          description: 선택 재료 여부
          default: false
        substitutes:
          type: array
          nullable: true
          items:
            type: string
          description: 대체 재료 목록

    CookingStep:
      type: object
      required:
        - step_number
        - instruction
      properties:
        step_number:
          type: integer
          description: 단계 번호
          example: 1
        instruction:
          type: string
          description: 조리 설명
        duration_seconds:
          type: integer
          nullable: true
          description: 소요 시간 (초)
        tips:
          type: string
          nullable: true
          description: 조리 팁
        image_url:
          type: string
          nullable: true
          description: 단계 이미지 URL

    Tag:
      type: object
      required:
        - name
        - category
      properties:
        name:
          type: string
          description: 태그명
          example: 한식
        category:
          type: string
          description: 카테고리
          enum: [cuisine, meal_type, dish_type, cooking_method, dietary]

    RecipeListResponse:
      type: object
      required:
        - items
        - has_more
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RecipeListItem'
        next_cursor:
          type: string
          nullable: true
          description: 다음 페이지 커서
        has_more:
          type: boolean
          description: 더 많은 결과 존재 여부

    # --- Chef Schemas ---
    ChefListItem:
      type: object
      required:
        - id
        - name
        - recipe_count
      properties:
        id:
          type: string
          format: uuid
          description: 요리사 UUID
        name:
          type: string
          description: 요리사 이름
          example: 백종원
        profile_image_url:
          type: string
          nullable: true
          description: 프로필 이미지 URL
        specialty:
          type: string
          nullable: true
          description: 전문 분야
          example: 한식
        recipe_count:
          type: integer
          description: 레시피 수
        is_verified:
          type: boolean
          description: 검증 여부
          default: false

    ChefSummary:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: 백종원
        profile_image_url:
          type: string
          nullable: true
        specialty:
          type: string
          nullable: true

    ChefDetail:
      allOf:
        - $ref: '#/components/schemas/ChefListItem'
        - type: object
          properties:
            bio:
              type: string
              nullable: true
              description: 자기 소개
            total_views:
              type: integer
              description: 총 조회수
            avg_rating:
              type: number
              format: float
              description: 평균 평점
            platforms:
              type: array
              items:
                $ref: '#/components/schemas/ChefPlatform'

    ChefPlatform:
      type: object
      required:
        - platform
        - platform_url
      properties:
        platform:
          type: string
          enum: [youtube, instagram, blog, naver, tiktok]
          description: 플랫폼 종류
        platform_url:
          type: string
          description: 플랫폼 URL
        subscriber_count:
          type: integer
          nullable: true
          description: 구독자/팔로워 수

    ChefListResponse:
      type: object
      required:
        - items
        - has_more
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ChefListItem'
        next_cursor:
          type: string
          nullable: true
          description: 다음 페이지 커서
        has_more:
          type: boolean
          description: 더 많은 결과 존재 여부

    # --- Error Schema ---
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: 오류 코드
              example: RECIPE_NOT_FOUND
            message:
              type: string
              description: 오류 메시지
              example: 레시피를 찾을 수 없습니다
            details:
              type: object
              additionalProperties: true
              description: 추가 오류 정보